pyproject.toml
[project]
name = "vigil"
version = "0.1.0"
description = "AI-powered, model-agnostic PR review tool"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "litellm>=1.0",
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
webhook = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]

[project.scripts]
vigil = "vigil.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vigil"]
